token qui marche + petit correction au remplacement de $
